requirements.txt
fastapi==0.99.0
uvicorn[standard]==0.23.1
python-multipart==0.0.6
Pillow==9.5.0
openai==0.27.0
python-dotenv==1.0.0
tinydb==4.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
